libs/cli/langchain_cli/integration_template/tests/integration_tests/test_chat_models.py
"""Test Chat__ModuleName__ chat model."""
from __module_name__.chat_models import Chat__ModuleName__


def test_stream() -> None:
    """Test streaming tokens from OpenAI."""
    llm = Chat__ModuleName__()

    for token in llm.stream("I'm Pickle Rick"):
        assert isinstance(token.content, str)


async def test_astream() -> None:
    """Test streaming tokens from OpenAI."""
    llm = Chat__ModuleName__()

    async for token in llm.astream("I'm Pickle Rick"):
        assert isinstance(token.content, str)


async def test_abatch() -> None:
    """Test streaming tokens from Chat__ModuleName__."""
    llm = Chat__ModuleName__()

    result = await llm.abatch(["I'm Pickle Rick", "I'm not Pickle Rick"])
    for token in result:
        assert isinstance(token.content, str)


async def test_abatch_tags() -> None:
    """Test batch tokens from Chat__ModuleName__."""
    llm = Chat__ModuleName__()

    result = await llm.abatch(
        ["I'm Pickle Rick", "I'm not Pickle Rick"], config={"tags": ["foo"]}
    )
    for token in result:
        assert isinstance(token.content, str)


def test_batch() -> None:
    """Test batch tokens from Chat__ModuleName__."""
    llm = Chat__ModuleName__()

    result = llm.batch(["I'm Pickle Rick", "I'm not Pickle Rick"])
    for token in result:
        assert isinstance(token.content, str)


async def test_ainvoke() -> None:
    """Test invoke tokens from Chat__ModuleName__."""
    llm = Chat__ModuleName__()

    result = await llm.ainvoke("I'm Pickle Rick", config={"tags": ["foo"]})
    assert isinstance(result.content, str)


def test_invoke() -> None:
    """Test invoke tokens from Chat__ModuleName__."""
    llm = Chat__ModuleName__()

    result = llm.invoke("I'm Pickle Rick", config=dict(tags=["foo"]))
    assert isinstance(result.content, str)
